requirements.txt
django-debug-toolbar>=0.9.4
django>=1.4
psycopg2>=2.4.5
South>=0.7.4
-e git+https://github.com/toastdriven/django-haystack.git@master#egg=django-haystack
requests
-e git+https://github.com/toastdriven/pyelasticsearch.git@master#egg=pyelasticsearch
django-tastypie>=0.9.11
